Introduce complex to Frequency Bin (#87)
* Introduce complex to Frequency Bin

* Adding a constructor for convenience

* Adding a constructor for convenience

* Restructured bin calculator

* cleanup

* cleanup

+++ src/main/kotlin/sound/bins/frequency/dominant/DominantBassFrequencyBinCalculator.kt
@@ -15,7 +15,8 @@ class DominantBassFrequencyBinCalculator(
         val frequency = dominantFrequencyCalculator.calculate(frequencyBins) ?: return null
         val magnitude = dominantMagnitudeCalculator.calculate(frequencyBins) ?: return null
 
-        return FrequencyBin(frequency, magnitude)
+
+        return FrequencyBin(frequency, magnitude.toDouble())
     }
 
 }
